requirements.txt
# ─────────────────────────────────────────
# IDP System – Backend Dependencies
# Tested & compatible set (Render-safe)
# ─────────────────────────────────────────

# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Async HTTP client (Gemini calls)
httpx==0.27.2

# File parsing
PyMuPDF==1.24.13          # PDF → fitz
python-docx==1.1.2         # .docx
openpyxl==3.1.5            # .xlsx engine
pandas==2.2.3              # .xlsx tabular data
python-pptx==1.0.2         # .pptx

# AI SDK
google-generativeai==0.8.3

# Utilities
python-multipart==0.0.12   # FastAPI file uploads
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
tiktoken==0.8.0            # Token-aware chunking
tenacity==9.0.0            # Retry logic
loguru==0.7.2              # Structured logging
